+++ nfl_predictor/utils/polars/finalize.py
@@ -62,6 +62,7 @@ def build_final_column_order() -> list[str]:
     base_stats.extend(pbp_stats)
     base_stats.extend(strength_stats)
     base_stats.extend(constants.QB_PBP_STATS)
+    base_stats.extend(constants.QB_SCHEDULE_STATS)
 
     # Deduplicate
     seen = set()